% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{plot_interactions}
\alias{plot_interactions}
\title{plot the interaction between two variables according to rf predictions}
\usage{
plot_interactions(final_rf, x, var1, var2, grid = 300, virtis_option = "D")
}
\description{
plot the interaction between two variables according to rf predictions
}
