% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_data.R
\name{read_multitab_NPX}
\alias{read_multitab_NPX}
\title{Multitab NPX data parsing}
\usage{
read_multitab_NPX(
  filename,
  sample_manifest = NULL,
  pheno = NULL,
  skip_mod = c(0),
  panel = c("NA"),
  num_tabs = 1,
  this_exp = "plasma"
)
}
\arguments{
\item{filename}{Name of the multitab excel raw data}

\item{sample_manifest}{Per individual pheno data}

\item{pheno}{Per sample pheno data}

\item{skip_mod}{The number of lines to skip in the excel file}

\item{panel}{Which panel is being extracted}

\item{num_tabs}{The number of tabs in the excel file}

\item{this_exp}{The sample matrix}
}
\description{
Multitab NPX data parsing
}
