% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{run_rf}
\alias{run_rf}
\title{run random forests}
\usage{
run_rf(
  long,
  variable = "grouped_severity",
  sampling = NULL,
  method = "rf",
  which_data = "protein"
)
}
\description{
run random forests
}
