% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mixed_model.R
\name{simple_mixed_de}
\alias{simple_mixed_de}
\title{simple mixed differential expression analysis plot}
\usage{
simple_mixed_de(
  long,
  case_control_to_remove = NA,
  variable = "case.control",
  covariates = NULL,
  random = c("Individual.ID"),
  plot_xlab = "Log2 Fold Change (P-N)",
  plot_title = "plasma case/control",
  labels = 0.1,
  logp_label = 10
)
}
\description{
simple mixed differential expression analysis plot
}
