% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mixed_model.R
\name{single_lmer}
\alias{single_lmer}
\title{run a single mixed model}
\usage{
single_lmer(data, formula_string)
}
\arguments{
\item{data}{The NPX data for a single protein}

\item{formula_string}{The formula of the model to be fit}
}
\description{
generates a model for a single protein. Attempts to use the default linear 
mixed model optimizer but switches to Nelder_Mead given an error
}
